requirements.txt
SpeechRecognition==3.10.4
pyttsx3==2.99
PyAudio==0.2.14
scikit-learn==1.5.2
pandas==2.2.3
numpy==2.1.2
requests==2.32.3
python-dotenv==1.0.1
customtkinter==5.2.2
google-api-python-client==2.151.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
dateparser==1.2.0
